### 2019-08-09 #1634 @whedon
[PRE REVIEW]: Kanapy: A Python package for generating complex synthetic polycrystalline microstructures.
**Submitting author:** @mrgprasad (<a href="http://orcid.org/0000-0003-2546-0161">Mahesh R. G. Prasad</a>)
**Repository:** <a href="https://github.com/mrgprasad/kanapy" target ="_blank">https://github.com/mrgprasad/kanapy</a>
**Version:** v0.0.1
