### 2019-08-09 #1634 @arfon
:wave: @mrgprasad - this PR should fix your paper: https://github.com/mrgprasad/kanapy/pull/1
