### 2019-08-24 #1634 @Kevin-Mattheus-Moerman
@heprom @ondrolexa @Mote3D would you be interested in reviewing this package for the [Journal of Open Source Software (JOSS)](https://joss.theoj.org/about). This submission is a Python package for generating complex synthetic microstructures. The review process focuses largely on [the software](https://github.com/mrgprasad/kanapy) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.01634/joss.01634/10.21105.joss.01634.pdf). Let me know if you have questions.
