### 2019-09-16 #1634 @putanowr
Hello,

I am sorry for the delayed answer. I have looked at the review guidelines as well as at the software and article itself.
I must say that I am very interested in you journal as the mission of it is something I also value much.

Regarding this particular review - after looking at all my duties and the fact that in two weeks I start new academic
year I have to say that I am available for review only after 20-th of October, when the this settle down a bit.
I am sorry if this is not the answer you expected, however delaying the review is the worse.

With kind regards,

Roman

September 12, 2019 4:51 PM, "Kevin Mattheus Moerman" <notifications@github.com (mailto:%22Kevin%20Mattheus%20Moerman%22%20<notifications@github.com>)> wrote:
	@putanowr (https://github.com/putanowr) We usually aim for ~3 weeks but we can be quite flexible. The review process is quite streamlined (and is often quite fast and easy to complete for reviewers) and everything takes place on GitHub. This here is a pre-review issue. Once I have sufficient reviewers I open a review issue. In the review issue each reviewer has a checklist which guides them through the review processes. Once all boxes are ticked we are ready to accept the work. The tickboxes in the list focus on the software functionality and the short paper. Below is more information.
	* Review guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html (https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
	* Example of a completed review issue for you to see the process: #1543 (https://github.com/openjournals/joss-reviews/issues/1543)
	Please can you confirm if you are able to help with this review? Thanks!
