### 2019-09-16 #1732 @Kevin-Mattheus-Moerman
@putanowr now that I have 3 reviewers it is okay if you are not able to join us. The review has now started. However if you are keen to help and the review process is still ongoing mid-October (which is when you said you might have time over at https://github.com/openjournals/joss-reviews/issues/1634), I can still add you here.
