### 2019-09-16 #1634 @putanowr
That's OK, please do so.

September 16, 2019 10:49 AM, "Kevin Mattheus Moerman" <notifications@github.com (mailto:%22Kevin%20Mattheus%20Moerman%22%20<notifications@github.com>)> wrote:
	@putanowr (https://github.com/putanowr) that is okay. Would you mind if I add you as reviewer anyway and we'll wait for your contribution mid-october? Thanks.
