### 2019-09-16 #1732 @putanowr
Well, if you add me as sort of backu-reviewer I understand I will be notified about the raised issues. This might be helpful
as I can get more familiar with the whole issue and will have more incentive to look how the journal operates.
Of course if that not disturb the normal process in any way.

Regard,

Roman

September 16, 2019 10:57 AM, "Kevin Mattheus Moerman" <notifications@github.com (mailto:%22Kevin%20Mattheus%20Moerman%22%20<notifications@github.com>)> wrote:
	@putanowr (https://github.com/putanowr) now that I have 3 reviewers it is okay if you are not able to join us. The review has now started. However if you are keen to help and the review process is still ongoing mid-October (which is when you said you might have time over at #1634 (https://github.com/openjournals/joss-reviews/issues/1634)), I can still add you here.
