### 2019-10-01 mrgprasad/kanapy#2 @KedoKudo
CLI improvements
@mrgprasad, the current version of `kanapy` seems to be missing a CLI interface that can take in a configuration file and perform the construction of a synthetic microstructure.  For such a CLI, I think the following features would be valuable for the micromechanics community, especially those using HPCC on a regular basis:

* A flexible and user friendly CLI built with established CLI libraries (click, docopt, etc.).
* Slightly more control over the generation of the orientations (preferred directions).
* Customizable workflow build on top of the existing modularized design of `kanapy`.
* The configuration file is in a popular format such as `yml`, `json`, `ini`, or`xml`.
