### 2019-10-01 mrgprasad/kanapy#3 @KedoKudo
Documentation improvements
@mrgprasad
Currently `kanapy` has a very extensive documentation, but I think most users could appreciate if the following aspects of the documentation can be improved:

* Brief explanation of the *.dump file content
* How to control the output unit in the exported file
* Benchmark of critical aspects of `kanapy`
* Change `src.kanapy.*` to `kanapy.*` in the documentation
