### 2019-10-01 mrgprasad/kanapy#4 @KedoKudo
Misc improvements
@mrgprasad
Here are some other improvement suggestions for `kanapy`:

* Please use `json` as the file extension for all JSON files instead of `txt` to avoid confusion.
* Some CLI that can generate a template config file for users to fill in the details.
* Use `tqdm` for progress bar.
* Provide a notebook example, and perhaps using Binder as well.
* Add coloring capability of individual grains (such as inverse pole figure coloring)
* Link to other synthetic microstructure generation tools available.
