### 2019-10-01 #1732 @KedoKudo
@Kevin-Mattheus-Moerman
After carefully reviewing `kanapy`, I think the following aspects of `kanapy` could be improved:

* Add a CLI to allow easy integration into existing workflow on a high performance computing environment (https://github.com/mrgprasad/kanapy/issues/2).
* The documentation can benefit from the features laid out in https://github.com/mrgprasad/kanapy/issues/3
* Some other minor improvements detailed in https://github.com/mrgprasad/kanapy/issues/4

The writing of the summary is adequate, but I hope the author could add a few sentences to compare `kanapy` with existing synthetic microstructure construction softwares (`Dream.3D` for example) to highlight the innovative aspect of `kanapy`.   Also, I noticed that the acknowledge part of `paper.md` is empty, which I am not sure whether this is intentional or not.
