### 2019-10-02 #1732 @mrgprasad
@KedoKudo
Thank you for taking time and reviewing the repo and the paper. I will make the necessary changes and get back to you soon.
