### 2019-10-02 #1732 @chennachaos
Hi @mrgprasad. I have started the review for `Kanapy`.
I have a remark related to **State of the field** in the **Software paper** section.

The paper discusses the issues with a couple of techniques used for generating synthetic microstructures, and then presents algorithms used in `Kanapy`, which is good. But, it does not mention anything about the need for Kanapy package. To convince the reader about Kanapy, it is important to address the following questions.
1.) What is motivation behind developing `Kanapy`?
2.) What are the advantages of `Kanapy` over other software libraries used for the generation of microstructures, for example, `Dream.3D` and `Neper`?
3.) Does `Kanapy` address any limitations of the existing libraries?
4.) Are there any specific features in `Kanapy` that are not available in other libraries? If so, what are they?
