### 2019-10-02 mrgprasad/kanapy#3 @mrgprasad
@KedoKudo
Thank you for the suggestions.
Could you please elaborate on your benchmark comment?
Do you mean benchmarking the particle packing simulation time?
