### 2019-10-02 mrgprasad/kanapy#4 @mrgprasad
@KedoKudo
Regarding the template configuration file - Do you mean the input file format should be changed from '.txt' to '.json' or '.yml' file formats ?

Regarding CLI that can generate config file - I don't clearly understand what this means, could you point me towards an existing implementation ?
