### 2019-10-02 #1732 @mrgprasad
@chennachaos
Thank you for the paper review and pointing out that I have skipped over the actual motivation for developing Kanapy. I will address this and the related questions and push a new version of the paper soon.
