### 2019-10-02 mrgprasad/kanapy#3 @KedoKudo
As a user, I would prefer to know what is the most time consuming aspect of `kanapy` and the corresponding scalability (lognorm, linear, quadratic, etc.).  I can use these information to extrapolate a rough estimate of the total computational time required for my workflow.  For example, if I am using `kanapy` to generate the microstructure input for my micromechanics simulation, I need to decide whether I want to bundle the two steps into one HPCC job.  If generating a microstructure only accounts for a fraction of the total simulation time and it scales very well, I would probably bundle two steps into one and use the total simulation time as a rough estimate to request time on HPCC.  Otherwise, I would most likely submit separate requests to maximize my chance of getting the time.
