### 2019-10-02 mrgprasad/kanapy#4 @KedoKudo
Running `kanapy` examples will generate a folder of _JSON files_ with _'txt'_ as file extension.  This is somewhat confusing as I was expecting an ascii table or some descriptive output.  I think it would be for the best if the file extension can reflect the actual content format.
Also, I would also recommend using a __single configuration file format__ (yml, json, etc.) for `kanapy` instead of using different ones at different locations, which only makes it harder for users to adopt `kanapy` into their workflow.
