### 2019-10-04 mrgprasad/kanapy#2 @mrgprasad
@KedoKudo
I just pushed an update to Kanapy's repo. Based on your suggestions, I have made the following changes:

1. A CLI is built using the library click for Kanapy. The **Usage** section of Kanapy's documentation has been updated in this regard.
2. The file extensions of input and the output files are all in json file format now.

**Regarding your orientations comment**: The user has complete control over the generations of orientations. The mean and sigma of the normal distribution are used as input parameters for this. (By orientations I mean the tilt angle of the ellipsoidal particles with respect to the positive x-axis and not the crystallographic orientations.)

**Regarding your comment on customizable workflow**: I have detailed this aspect in the **Usage** section of the documentation.
