### 2019-10-04 mrgprasad/kanapy#3 @mrgprasad
@KedoKudo
Based on you suggestions for documentation improvement, I have updated the following:

1. The contents of the .dump file has been described in the **Usage** section under the heading: *Visualize the packing simulation*.

2. The user can now specify in the input file, the **output unit** in which the ABAQUS file and the Output statistics can be exported.

3. In the documentation, under the section **Applications**, I have added a sub-section called **Simulation benchmarks**. This details the CPU time required for particle packing and voxelization routines. These are the critical aspects of Kanapy.

4. *src.kanapy.* has been changed to *kanapy.* in the documentation.
