### 2019-10-04 mrgprasad/kanapy#4 @mrgprasad
@KedoKudo
Based on you suggestions for Misc improvement, I have made the following changes in Kanapy:

1. All files are now in *json* file format, except for the files outputted to Neper for tessellation.

2. *tqdm* has now been used in place of the previously used progress bar.

3. **Regarding your notebook example (binder) comment**: In the future we plan to include crystallographic texture modeling, to make Kanapy a complete microstructure generation tool. The work is ongoing and requires MTex (a Matlab library). We are sceptical about the interactions between Binder's notebooks and matlab functions. Hence we don't intend to provide the notebook (Binder) functionality for Kanapy.

4. Once we have the crystallographic texture part included in Kanapy, we can provide a unique color to each particle based on its orientation. As of now, the visualization software OVITO can be used for color coding the particles.

5. A link to Neper in the form of particle positions and weights files is already provided by Kanapy for spherical grains. In the future, we intend to provide a link to Dream.3D, but this plan can be executed once texture is also included.
