### 2019-10-04 #1732 @mrgprasad
@Kevin-Mattheus-Moerman, @KedoKudo, @chennachaos and @heprom
I have pushed an update to the Kanapy repo based on KedoKudo's  review.
