### 2019-10-04 #1732 @mrgprasad
@KedoKudo
I have updated the software paper based on your comments. Do let me know if you have any further suggestions. Also, I have removed the acknowledgement section.

@chennachaos
With this update I have also tried to address our queries. I hope the motivation for Kanapy's development, its design and unique features and how it addresses the limitations of other software libraries is clear now.
