### 2019-10-09 mrgprasad/kanapy#4 @KedoKudo
@mrgprasad
Thank you for making the improvements and I am looking forward to the texture capability of _kanapy_.  I will close this issue now.
