### 2019-10-09 mrgprasad/kanapy#2 @KedoKudo
@mrgprasad
Thank you for taking the time making these improvements and adding the corresponding documentation.  As for the _orientation_ comment, most people in the metallurgy field would refer the term _orientation_ to crystallographic orientation as this is important for a lot of mesoscale analysis.   Since the synthetic texture information will be added in the future release of _kanapy_ according to #4, I believe I can safely close this issue to move forward.
