### 2019-10-09 mrgprasad/kanapy#3 @KedoKudo
@mrgprasad
Thank you for adding the benchmark and various other edits in the documentation.  I will close this issue now.
