### 2019-10-09 #1732 @KedoKudo
@mrgprasad @Kevin-Mattheus-Moerman
I am happy with the current state of `kanapy`, and I would like to recommend this package for publication.
