### 2019-10-10 #1732 @chennachaos
@mrgprasad
**A statement of need** for *Documentation* is missing.

Consider adding a paragraph on 'motivation' behind Kanapy in the *Overview* section of the documentation.

I will open these two as issues in your Github repo so we can track them.
