### 2019-10-10 #1732 @mrgprasad
@chennachaos
I have pushed an update to Kanapy's repo based on your review. I have also updated the issues you have opened on Kanapy's repo. Do let me know if additional changes are necessary.
