### 2019-10-11 #1732 @chennachaos
Hi @Kevin-Mattheus-Moerman, I have completed the review. @mrgprasad has been prompt in responding to my questions and remarks. I am happy with the current state of the `Kanapy` package.

I recommend the `Kanapy` software package for publication in the JOSS.

Thank you for the opportunity.
