### 2019-10-19 #1732 @heprom
Hi @Kevin-Mattheus-Moerman I have done a first pass at reading the paper and testing the code. There is a small bug in the examples (I thing after moving to json input files): the input file in ellipsoid_pack.py has not been changed which result in an error when running the example.
