### 2019-10-19 #1732 @heprom
A comment on the paper: first it is concise and well written. I recommend changing from "With the advent of additive manufacturing, the processing steps usually result in rather complex microstructures with elongated grains and strong crystallographic textures in metals." to something less reductive like: "Metallic material processing (including additive manufacturing) often result in complex microstructures with elongated grains and strong crystallographic textures."
