### 2019-10-19 #1732 @mrgprasad
@heprom
Thank you for the software and paper review. I have pushed an update to Kanapy's repo.
With this update I have fixed the bug in **ellipsoid_pack.py** file and the corrected the above mentioned sentence in the JOSS paper.
