### 2019-10-19 #1732 @heprom
yes I'm happy with the software now.

On Sat, Oct 19, 2019 at 10:13 PM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @heprom <https://github.com/heprom> thanks for your review comments! Do
> you feel @mrgprasad <https://github.com/mrgprasad> has addressed the
> issues you raised?
>
