### 2019-10-20 #1732 @heprom
another small detail in the examples (both):
shouldn't the path to the input file be defined as

`inputFile = os.path.join(os.getcwd(), 'stat_input.json')`

so that it is plateform independent. I'm running linux so it works for me as is but I guess this will not work on windows.
