### 2019-10-20 #1732 @mrgprasad
@heprom
Currently we intend provide Kanapy for Linux & MAC OS platforms only.
In the future, with additional functionalities (that is still under development) we may release it for Windows too.
