### 2019-10-21 #1732 @mrgprasad
@Kevin-Mattheus-Moerman
Please let me know if anything more is required from my side.

For a future release of Kanapy, we are planning to include functionalities for modeling crystallographic texture. This would be a major update for the software. In this regard:

* Is it possible to submit the software for review again?
* If not, can we update the JOSS paper somehow?
