### 2019-11-04 #1732 @Kevin-Mattheus-Moerman
- [ ] Your paper is about to be processed for acceptance. Please double check your paper and ensure you are happy with the author list and acknowledgements as well.
- [ ] Can you work on those missing DOI's?
- [ ] Can you check this citation: `(Groeber & Jackson, 2014, p. @Vajragupta2014)`? The `p.` seems you indicate your citing a page but the number is missing?
- [ ] Check citation syntax for: `Wang et al. (Wang, Wang, & Kim, 2001)`, see also: https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax
- [ ] I suggest removal of the space in workflows in `...to create individual work flows for generating...`.
- [ ] Can  you add citations to SciPy and NumPy? in `...from the Scipy package together with various Numpy array operations...`
