### 2019-11-04 #1732 @mrgprasad
@Kevin-Mattheus-Moerman
With the latest push to the repo, I have made the necessary changes for the paper. DOI's don't exist for scipy, numpy, pybind11 and eigen libraries. Please let me know if additional changes are required.
