### 2019-11-05 #1732 @Kevin-Mattheus-Moerman
@mrgprasad thanks for implementing most changes. I left the box in relation to the citation issue unchecked.

Can you check the citation information presented here:
https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax

In your paper you mention `....developed by Wang et al. (Wang, Wang, & Kim, 2001).`. This needs to be altered. I suggest that instead of `...developed by Wang et al. [@Wang2001]` which you currently have, you use `...developed by Wang et al. [-@Wang2001]`, the added minus sign `-` will suppress the second appearance of the author name(s) since  you already mention it in the text.
