### 2019-11-05 #1732 @Kevin-Mattheus-Moerman
@mrgprasad great. Thanks looks good to me now. Can you now do the following:
- [ ] Archive a copy of the reviewed software on Zenodo and list the DOI of the archived version here. Make sure that as you do this that the meta data (e.g. the title and author list) will match that of the paper.
- [ ] Can you please confirm the version of the reviewed and archived software? Is it still `v0.0.1` or have you changed the version?
