### 2019-11-05 #1732 @Kevin-Mattheus-Moerman
@openjournals/joss-eics I recommend this submission for acceptance in JOSS
