### 2019-11-08 #1732 @kyleniemeyer
Hi @mrgprasad, just a few fixes needed in the paper:
- missing period in i.e in second paragraph: "i.e," -> "i.e.,"
- in the 3rd paragraph, the sentence beginning with "Thus addressing" is incomplete, or at least awkward. Perhaps something like "This addresses the limitations..."
- last sentence of 3rd paragraph: change "in-built" to "built-in"
- pybind11 citation:
```
@misc{pybind11,
   author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
   year = {2017},
   howpublished = {\url{https://github.com/pybind/pybind11}},
   title = {pybind11 -- Seamless operability between {C++11} and {Python}}
}
```

Please let me know back here when you've made those changes and I'll finish the last few steps of acceptance.
