### 2019-11-08 #1732 @mrgprasad
@kyleniemeyer
I have made all the changes based on your suggestions. Thank you.
