### 2019-11-08 #1732 @mrgprasad
I would like to thank @KedoKudo, @chennachaos & @heprom for reviewing Kanapy. Your suggestions have made the software better.
@Kevin-Mattheus-Moerman thank you for taking time and editing!
And final thanks to @kyleniemeyer for accepting the submission.
