### 2019-11-08 #1732 @mrgprasad
@kyleniemeyer
We are working on extending Kanapy to include other important features with respect to microstructure generation. In this regard, can we submit the software for review again? If not, is it possible to update the current JOSS paper to reflect this?
