### 2019-11-08 #1732 @kyleniemeyer
@mrgprasad if the software functionality changes substantially, then you can submit later for a new review.
